Revert "move home to pages foldet"
This reverts commit 182539b4e846fd000ee1523d447a0dc716e7ca0f.
